requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
tensorflow==2.15.1
matplotlib==3.7.1
Pillow==9.4.0